requirements/development.txt
aiosmtplib==2.0.1
alembic==1.10.3
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
blinker==1.6
certifi==2022.12.7
click==8.1.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.95.0
fastapi-mail==1.2.7
fastapi-sso==0.6.4
fastapi-utils==0.2.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
oauthlib==3.2.2
orjson==3.8.9
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg-pool==3.1.7
psycopg2-binary==2.9.6
pyasn1==0.4.8
pydantic==1.10.7
pytest==7.2.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
redis==4.5.4
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
sqlparse==0.4.4
sqltap==0.3.11
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
Werkzeug==2.2.3
